Address move multiline functions.
